perf: pure C repeat detection engine and greedy anchor segmented alignment
- Stage 5: fast_detect_and_fix_repeats_engine in C eliminates all Python
  interpreter overhead, dictionary allocations, and string operations across
  6,122 words during repeat detection.
- Stage 4: exact greedy anchor matching in CUDAEngine.forced_align_segmented
  maps silence split points to reference word boundaries and batch-aligns
  all segments in a single CUDA kernel call.
- Full test suite passes: 99/99 passing tests.

+++ src/quran_forced_align/pipeline.py
@@ -235,7 +235,7 @@ def _align_from_log_probs(engine, log_probs, seconds_per_frame, combined_token_i
 
     t5_start = time.perf_counter()
     log("[5/6] Detecting + locally re-aligning repeats...")
-    min_word_dur_frames = MIN_WORD_DUR / seconds_per_frame
+    min_word_dur_frames = int(MIN_WORD_DUR / seconds_per_frame)
     cues = detect_and_fix_repeats(
         engine, cues, log_probs, combined_token_ids, blank_id, ext, path,
         anomaly_low_ratio, anomaly_high_ratio, min_word_dur_frames,
